pyproject.toml
[project]
name = "fastapi-render"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.1",
    "langchain-google-genai>=3.0.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
]
